Update profile and registration Screen

+++ app/src/main/java/com/example/testonlinestore/domain/model/registration/UserAccount.kt
@@ -2,6 +2,7 @@ package com.example.testonlinestore.domain.model.registration
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.util.UUID
 
 
 @Entity(tableName = "user_account")
@@ -9,5 +10,5 @@ data class UserAccount(
     val name : String,
     val surname : String,
     @PrimaryKey(autoGenerate = true)
-    val number : Int
+    val number : Long
 )
